+++ src/dsp/src/mixer.c
@@ -88,13 +88,12 @@ gkick_mixer_process(struct gkick_mixer *mixer,
         for (size_t channel = 0; channel < GEONKICK_MAX_CHANNELS; channel++) {
                 size_t left_index  = 2 * channel;
                 size_t right_index = left_index + 1;
-                for (size_t i = 0; i < GEONKICK_MAX_INSTRUMENTS; i++) {
+                for (size_t i = 0; i < GEONKICK_MAX_INSTRUMENTS + 1; i++) {
                         struct gkick_audio_output *output = mixer->audio_outputs[i];
                         if (output->channel != channel)
                                 continue;
 
                         if (!output->enabled || output->muted || mixer->solo != output->solo) {
-                                ring_buffer_reset(output->ring_buffer);
                                 output->play = false;
                         } else {
                                 if (output->play)
@@ -105,34 +104,18 @@ gkick_mixer_process(struct gkick_mixer *mixer,
                                 ring_buffer_get_data(output->ring_buffer,
                                                      out[right_index] + offset,
                                                      size);
-                                gkick_real leveler_val = ring_buffer_get_cur_data(output->ring_buffer);
-                                ring_buffer_next(output->ring_buffer, size);
-
                                 gkick_real limiter_val = (gkick_real)output->limiter / 1000000;
                                 gkick_mixer_apply_limiter(out[left_index] + offset,
                                                           out[right_index] + offset,
                                                           size,
                                                           limiter_val);
+                                gkick_real leveler_val = ring_buffer_get_cur_data(output->ring_buffer);
+                                leveler_val *= limiter_val;
                                 gkick_mixer_set_leveler(mixer, i, leveler_val);
                         }
+                        ring_buffer_next(output->ring_buffer, size);
                 }
         }
-
-        struct gkick_audio_output *output = mixer->audio_outputs[GEONKICK_MAX_INSTRUMENTS];
-        if (output->play)
-                gkick_audio_set_play(output);
-
-        ring_buffer_get_data(output->ring_buffer,
-                             out[0] + offset,
-                             size);
-        ring_buffer_get_data(output->ring_buffer,
-                             out[1] + offset,
-                             size);
-        ring_buffer_next(output->ring_buffer, size);
-        gkick_mixer_apply_limiter(out[0] + offset,
-                                  out[1] + offset,
-                                  size,
-                                  (gkick_real)output->limiter / 1000000);
 
         return GEONKICK_OK;
 }